tests: organise the suite so a finding can be committed before its fix
Preparation for a review round whose output is failing tests rather than prose.
Three things stood in the way.

There was no way to commit a reproduction without also committing the fix, so a
defect found and not immediately fixed survived only as a paragraph somebody had
to believe. Runner.Pending takes a test that is EXPECTED to fail and a reason:
it reports PEND while it still fails and keeps the run green, and it FAILS the
run the moment it starts passing, saying that the defect is fixed and this should
become an ordinary Test. The inversion is the point - an expected-failure marker
that can rot into a permanently ignored test is worse than no marker. Verified
both directions before use.

Placement was ambiguous exactly where the work is. TLS exists on two separate
stacks and three files shared a name across projects, so "where does my rotation
test go" had no answer from the tree. The QUIC-side files are now
QuicMutualTlsTests, QuicSniTests and QuicRotationTests, and the README opens with
a table that answers the question directly.

And the files were dense enough that a dozen people editing them would collide.
The suite is split along seams that already existed: configuration refusals out
of MutualTlsTests and SniTests, the validating-client rotation tests out of
RotationTests, and the 1228-line TestServer.cs into TestServer, Client and
Externals. No file now holds more than 14 tests.

Splitting surfaced duplication worth removing: the same send-first TLS handler
was copied into two suites verbatim, and two different handlers were both called
IdentityHandler. They are now Handlers.TlsSendFirst, TlsIdentity, TlsCommonName
and TlsIdentitySendFirst, in the shared harness.

The README is now a contributor guide rather than an inventory: where a test
goes, prove it fails first, assert why and not that, guard against passing
vacuously, pair a negative with a control, never assert on timing, and how to
mint the awkward certificates - including the two fixture traps that cost real
time (a leaf cannot predate its issuer, and an EC leaf cannot be signed through
Create(issuer, ...)).

302 tests green, unchanged in count: this moves and deduplicates, it does not
alter behaviour.
