+++ tests/Ioxide.Tests.Tls/MutualTlsTests.cs
@@ -26,7 +26,7 @@ internal static class MutualTlsTests
                 (string ca, string serverCert, string serverKey, string clientCert, string clientKey, _, _)
                     = TestCert.EnsureMutualTls();
 
-                int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = serverCert,
                     KeyPath = serverKey,
@@ -45,7 +45,7 @@ internal static class MutualTlsTests
             {
                 (string ca, string serverCert, string serverKey, _, _, _, _) = TestCert.EnsureMutualTls();
 
-                int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = serverCert,
                     KeyPath = serverKey,
@@ -63,7 +63,7 @@ internal static class MutualTlsTests
                 (string ca, string serverCert, string serverKey, _, _, string rogueCert, string rogueKey)
                     = TestCert.EnsureMutualTls();
 
-                int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = serverCert,
                     KeyPath = serverKey,
@@ -85,7 +85,7 @@ internal static class MutualTlsTests
 
             // The mixed port: anyone may connect, and the handler decides what an unauthenticated
             // one is allowed to reach.
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -104,7 +104,7 @@ internal static class MutualTlsTests
             (string ca, string serverCert, string serverKey, string clientCert, string clientKey, _, _)
                 = TestCert.EnsureMutualTls();
 
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -125,7 +125,7 @@ internal static class MutualTlsTests
 
             // "Optional" governs presenting NOTHING. A certificate that is offered is always
             // verified - otherwise the option would read as "trust anything".
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -142,7 +142,7 @@ internal static class MutualTlsTests
             (string ca, string serverCert, string serverKey, string clientCert, string clientKey, _, _)
                 = TestCert.EnsureMutualTls();
 
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -163,7 +163,7 @@ internal static class MutualTlsTests
 
             // No anchors configured: nothing is requested, so a client that HAS a certificate is
             // never asked for it and connects anonymously.
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = certPath,
                 KeyPath = keyPath,
@@ -176,7 +176,6 @@ internal static class MutualTlsTests
         });
 
         RegisterBadClientCertificates(runner);
-        RegisterConfigurationErrors(runner);
     }
 
     // Configuration mistakes that should fail where they are written, not at some later handshake.
@@ -220,7 +219,7 @@ internal static class MutualTlsTests
                 (string ca, string cert, string key) = TestCert.EnsureClientCert(spec);
                 (string serverCert, string serverKey) = ServerFor(ca);
 
-                int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsIdentity, r => TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = serverCert,
                     KeyPath = serverKey,
@@ -242,7 +241,7 @@ internal static class MutualTlsTests
             (string anchors, _, string cert, string key) = TestCert.EnsureChainedClientCert();
             (string serverCert, string serverKey) = ServerFor(anchors);
 
-            int port = TestServer.Start(CommonNameHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsCommonName, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -266,7 +265,7 @@ internal static class MutualTlsTests
             });
             (string serverCert, string serverKey) = ServerFor(ca);
 
-            int port = TestServer.Start(CommonNameHandler, r => TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsCommonName, r => TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -287,137 +286,6 @@ internal static class MutualTlsTests
         return (serverCert, serverKey);
     }
 
-    private static void RegisterConfigurationErrors(Runner runner)
-    {
-        runner.Test("mtls: a trust bundle with a corrupt block is refused, not silently truncated", () =>
-        {
-            // ClientCaPath and ClientCaPem are documented as equivalent in what they trust. They
-            // were not: the file route refuses such a bundle outright, while the PEM-text route
-            // read until it stopped and kept whatever came BEFORE the bad block - so every client
-            // issued by a later anchor was refused, with nothing said server-side. Failing closed
-            // is not the same as failing loudly, and an operator has no way to see the difference.
-            (string ca, _, _, _, _, _, _) = TestCert.EnsureMutualTls();
-
-            Assert.True(StartFails(new TlsOptions
-            {
-                CertificatePath = TestCert.Ensure().CertPath,
-                KeyPath = TestCert.Ensure().KeyPath,
-                ClientCaPem = BundleWithCorruptMiddle(ca),
-                RequireClientCertificate = true,
-            }, "malformed"), "a bundle with a corrupt block should be refused at startup");
-        });
-
-        runner.Test("mtls: requiring a certificate without anchors is refused", () =>
-        {
-            (string certPath, string keyPath) = TestCert.Ensure();
-
-            // Nothing to validate against: this would refuse every client sending none and accept
-            // anything from one that sends something - the opposite of what it reads as.
-            Assert.True(StartFails(new TlsOptions
-            {
-                CertificatePath = certPath,
-                KeyPath = keyPath,
-                RequireClientCertificate = true,
-            }, "RequireClientCertificate needs trust anchors"),
-                "RequireClientCertificate without anchors must be rejected");
-        });
-
-        runner.Test("mtls: two client CA sources are refused", () =>
-        {
-            (string ca, string serverCert, string serverKey, _, _, _, _) = TestCert.EnsureMutualTls();
-
-            Assert.True(StartFails(new TlsOptions
-            {
-                CertificatePath = serverCert,
-                KeyPath = serverKey,
-                ClientCaPath = ca,
-                ClientCaPem = File.ReadAllText(ca),
-            }, "At most one client CA source"),
-                "ClientCaPath and ClientCaPem together must be rejected");
-        });
-
-        runner.Test("mtls: the common name is reported exactly, not as part of a rendered subject", () =>
-        {
-            (string ca, string serverCert, string serverKey, string clientCert, string clientKey, _, _)
-                = TestCert.EnsureMutualTls();
-
-            int port = TestServer.Start(CommonNameHandler, r => TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = serverCert,
-                KeyPath = serverKey,
-                ClientCaPath = ca,
-                RequireClientCertificate = true,
-            }));
-
-            (int status, string body) = Client.GetTlsClientCert(port, "/who", clientCert, clientKey);
-
-            Assert.Equal(200, status);
-            // Equality, not Contains: the value is meant to be compared whole.
-            Assert.Equal("alice", body);
-        });
-
-        runner.Test("mtls: a subject crafted to look like another identity does not become one", () =>
-        {
-            // The reason PeerCommonName exists. The rendered subject escapes a literal '/' as
-            // "\\/", which still CONTAINS '/', so this legitimately-issued certificate renders as
-            //     /O=Acme\\/CN=admin.internal/CN=mallory
-            // and satisfies Contains("/CN=admin.internal") while being an entirely different
-            // principal. The structural CN is unmoved by any of that.
-            (_, string serverCert, string serverKey, _, _, _, _) = TestCert.EnsureMutualTls();
-            (string ca, string certPath, string keyPath) =
-                TestCert.EnsureClientCertFromCa("O=Acme/CN=admin.internal, CN=mallory");
-
-            int port = TestServer.Start(CommonNameHandler, r => TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = serverCert,
-                KeyPath = serverKey,
-                ClientCaPath = ca,
-                RequireClientCertificate = true,
-            }));
-
-            (int status, string body) = Client.GetTlsClientCert(port, "/who", certPath, keyPath);
-
-            Assert.Equal(200, status);
-            Assert.Equal("mallory", body);
-        });
-
-        runner.Test("mtls: the rendered subject is the one that IS confusable", () =>
-        {
-            // The other half of the pair, asserted so the trap is a recorded fact rather than a
-            // remark in a doc comment. If a future change makes the rendered form unambiguous this
-            // test fails, and that is the moment to revisit what the docs promise.
-            (_, string serverCert, string serverKey, _, _, _, _) = TestCert.EnsureMutualTls();
-            (string ca, string certPath, string keyPath) =
-                TestCert.EnsureClientCertFromCa("O=Acme/CN=admin.internal, CN=mallory");
-
-            int port = TestServer.Start(IdentityHandler, r => TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = serverCert,
-                KeyPath = serverKey,
-                ClientCaPath = ca,
-                RequireClientCertificate = true,
-            }));
-
-            (int status, string body) = Client.GetTlsClientCert(port, "/who", certPath, keyPath);
-
-            Assert.Equal(200, status);
-            Assert.True(body.Contains("/CN=admin.internal"),
-                $"expected the rendered subject to be substring-confusable, got: {body}");
-        });
-
-        runner.Test("mtls: an unreadable client CA fails at startup", () =>
-        {
-            (string certPath, string keyPath) = TestCert.Ensure();
-
-            Assert.True(StartFails(new TlsOptions
-            {
-                CertificatePath = certPath,
-                KeyPath = keyPath,
-                ClientCaPath = "/nonexistent/ca.pem",
-            }, "/nonexistent/ca.pem"),
-                "a missing CA file must be reported at startup");
-        });
-    }
 
     /// <summary>
     /// A trust bundle whose MIDDLE block is corrupt: a real anchor, then a well-formed PEM envelope
@@ -481,45 +349,6 @@ internal static class MutualTlsTests
 
     // Answers with the authenticated identity, or "anonymous". Reading it is the whole point of
     // the feature: a server that can only enforce cannot authorise.
-    private static Task IdentityHandler(Reactor reactor, TcpConnection connection)
-        => Identity(reactor, connection, static s => s.PeerSubject);
-
-    // The same, reporting the structural CN - the value an application is meant to authorize on.
-    private static Task CommonNameHandler(Reactor reactor, TcpConnection connection)
-        => Identity(reactor, connection, static s => s.PeerCommonName);
-
-    private static async Task Identity(Reactor reactor, TcpConnection connection,
-        Func<TlsSession, string?> report)
-    {
-        TlsSession? session = null;
-        try
-        {
-            session = await reactor.GetService<TlsService>()!.AcceptAsync(connection);
-
-            while (true)
-            {
-                RecvSnapshot snapshot = await connection.ReadAsync();
-                if (snapshot.IsClosed)
-                {
-                    return;
-                }
-
-                string subject = report(session) ?? "anonymous";
-                byte[] body = Encoding.ASCII.GetBytes(subject);
-
-                session.Write(connection, Encoding.ASCII.GetBytes(
-                    $"HTTP/1.1 200 OK\r\ncontent-length: {body.Length}\r\n\r\n{subject}"));
-
-                await connection.FlushAsync();
-                connection.ResetRead();
-            }
-        }
-        finally
-        {
-            session?.Dispose();
-            connection.DecRef();
-        }
-    }
 
     private static Task EmptyHandler(Reactor reactor, TcpConnection connection)
     {
