+++ tests/Ioxide.Tests.Tls/RotationTests.cs
@@ -26,7 +26,7 @@ internal static class RotationTests
                 (string second, string secondKey) = TestCert.EnsureNamed("second.test");
 
                 TlsService? service = null;
-                int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = first,
                     KeyPath = firstKey,
@@ -53,7 +53,7 @@ internal static class RotationTests
                 (string second, string secondKey) = TestCert.EnsureNamed("second.test");
 
                 TlsService? service = null;
-                int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+                int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
                 {
                     CertificatePath = cert,
                     KeyPath = key,
@@ -77,7 +77,7 @@ internal static class RotationTests
             (string beta, string betaKey) = TestCert.EnsureNamed("beta.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -110,7 +110,7 @@ internal static class RotationTests
             (string alpha, string alphaKey) = TestCert.EnsureNamed("alpha.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -137,7 +137,7 @@ internal static class RotationTests
             (string alpha, string alphaKey) = TestCert.EnsureNamed("alpha.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -164,7 +164,7 @@ internal static class RotationTests
             (string cert, string key) = TestCert.Ensure();
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -194,7 +194,7 @@ internal static class RotationTests
             (string alpha, string alphaKey) = TestCert.EnsureNamed("alpha.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -338,7 +338,7 @@ internal static class RotationTests
             (string beta, string betaKey) = TestCert.EnsureNamed("beta.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = serverCert,
                 KeyPath = serverKey,
@@ -375,7 +375,7 @@ internal static class RotationTests
             (string beta, string betaKey) = TestCert.EnsureNamed("beta.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = cert,
                 KeyPath = key,
@@ -416,7 +416,6 @@ internal static class RotationTests
                 $"the service should still serve a certificate for the name, got: {subject}");
         });
 
-        RegisterCurlRotation(runner);
 
         runner.Test("rotate: a validating client is served while rotations run underneath it", () =>
         {
@@ -429,7 +428,7 @@ internal static class RotationTests
             (_, string second, string secondKey) = TestCert.EnsureRenewedFromCa("rotate.test");
 
             TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
+            int port = TestServer.Start(Handlers.TlsSendFirst, r => service = TlsService.Start(r, new TlsOptions
             {
                 CertificatePath = first,
                 KeyPath = firstKey,
@@ -473,169 +472,8 @@ internal static class RotationTests
 
     }
 
-    private static void RegisterCurlRotation(Runner runner)
-    {
-        runner.Test("control: a validating client is served repeatedly with NO rotation", () =>
-        {
-            // The control, and it earns its place: when the rotation tests below first flaked, this
-            // failed too - which said the problem was the suite's CPU contention rather than
-            // anything about replacing certificates. Keep it, so the next person gets that answer
-            // in one run instead of an afternoon.
-            (string ca, string first, string firstKey) = TestCert.EnsureNamedFromCa("rotate.test");
 
-            int port = TestServer.Start(OkHandler, r => TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = first,
-                KeyPath = firstKey,
-            }));
 
-            for (int i = 0; i < 4; i++)
-            {
-                (int exit, string body, string stderr) = Curl.Get("rotate.test", port, ca);
-                Assert.True(exit == 0, $"request {i} without any rotation failed: {stderr}");
-                Assert.True(body.Contains("ok"), $"request {i} was not answered: {body}");
-            }
-        }, skip: !Curl.Available);
-
-        runner.Test("rotate: a validating client is served correctly across repeated rotations", () =>
-        {
-            (string ca, string first, string firstKey) = TestCert.EnsureNamedFromCa("rotate.test");
-            (_, string second, string secondKey) = TestCert.EnsureRenewedFromCa("rotate.test");
-
-            TlsService? service = null;
-            int port = TestServer.Start(OkHandler, r => service = TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = first,
-                KeyPath = firstKey,
-            }));
-
-            // Rotate, then ask - so every single request is served by a set that was installed a
-            // moment earlier, and any state a rotation leaves behind is what answers it.
-            // Five is enough: each iteration installs a whole new set and the request that follows
-            // is served by it. More only lengthens the window in which this suite's other reactors
-            // can starve the request, which is a property of the suite and not of rotation.
-            for (int i = 0; i < 5; i++)
-            {
-                bool even = i % 2 == 0;
-
-                service!.ReplaceCertificates(new TlsCertificate
-                {
-                    CertificatePath = even ? second : first,
-                    KeyPath = even ? secondKey : firstKey,
-                });
-
-                (int exit, string body, string stderr) = Curl.Get("rotate.test", port, ca);
-
-                Assert.True(exit == 0, $"request {i} after a rotation was refused: {stderr}");
-                Assert.True(body.Contains("ok"), $"request {i} after a rotation was not answered: {body}");
-            }
-        }, skip: !Curl.Available);
-
-        // A curl-driven CONCURRENT rotation test used to live here and was removed: with every
-        // test server in the suite still busy-polling, a rotator running underneath the requests
-        // only ever measured CPU contention, and failed as a timeout that reads like a server bug.
-        // Concurrency is covered in-process by the load test further down, which is far cheaper.
-
-        runner.Test("rotate: a client certificate is still verified across repeated rotations", () =>
-        {
-            // Both halves at once. The server's certificate is renewed before every request; the
-            // client's is checked on every one. Renewing must not disturb who may connect - and the
-            // impostor must stay out throughout, or a rotation is a window with no verification.
-            (string ca, string serverCert, string serverKey, string clientCert, string clientKey,
-             string rogueCert, string rogueKey) = TestCert.EnsureMutualTls();
-            (_, string renewed, string renewedKey) = TestCert.EnsureRenewedFromCa("localhost");
-
-            TlsService? service = null;
-            int port = TestServer.Start(IdentityHandler, r => service = TlsService.Start(r, new TlsOptions
-            {
-                CertificatePath = serverCert,
-                KeyPath = serverKey,
-                ClientCaPath = ca,
-                RequireClientCertificate = true,
-            }));
-
-            for (int i = 0; i < 4; i++)
-            {
-                bool even = i % 2 == 0;
-
-                service!.ReplaceCertificates(new TlsCertificate
-                {
-                    CertificatePath = even ? renewed : serverCert,
-                    KeyPath = even ? renewedKey : serverKey,
-                });
-
-                (int exit, string body, string stderr) = Curl.Get("localhost", port, ca, "/who", clientCert, clientKey);
-
-                Assert.True(exit == 0, $"the trusted client was refused after rotation {i}: {stderr}");
-                Assert.True(body.Contains("alice"),
-                    $"the trusted client should still be named after rotation {i}, got: {body}");
-
-                (int rogueExit, _, _) = Curl.Get("localhost", port, ca, "/who", rogueCert, rogueKey);
-
-                Assert.True(rogueExit != 0,
-                    $"an untrusted client got in after rotation {i} - a rotation must never be a window with no client verification");
-            }
-        }, skip: !Curl.Available);
-    }
-
-    /// <summary>Answers with the client's verified identity, so authentication can be seen.</summary>
-    private static async Task IdentityHandler(Reactor reactor, TcpConnection connection)
-    {
-        TlsSession? session = null;
-
-        try
-        {
-            session = await reactor.GetService<TlsService>()!.AcceptAsync(connection);
-
-            // The first request routinely rides in with the handshake's final flight - a TLS 1.3
-            // client sends it immediately after Finished, and those bytes are already decrypted and
-            // gone from the socket by the time AcceptAsync returns. Answering it BEFORE parking on
-            // a read is what the send-first loop means; waiting first hangs, because the bytes
-            // being waited for already arrived.
-            if (!session.DrainPlaintext().IsEmpty)
-            {
-                session.Write(connection, Identity(session));
-                await connection.FlushAsync();
-            }
-
-            // No ResetRead above: that belongs to a read that was actually issued, and calling it
-            // for one that never happened leaves the connection's read state describing a read
-            // nobody made.
-            while (true)
-            {
-                RecvSnapshot snapshot = await connection.ReadAsync();
-
-                if (snapshot.IsClosed)
-                {
-                    return;
-                }
-
-                byte[] who = System.Text.Encoding.ASCII.GetBytes(session.PeerSubject ?? "anonymous");
-                byte[] head = System.Text.Encoding.ASCII.GetBytes(
-                    $"HTTP/1.1 200 OK\r\ncontent-length: {who.Length}\r\n\r\n");
-
-                session.Write(connection, [.. head, .. who]);
-
-                await connection.FlushAsync();
-                connection.ResetRead();
-            }
-        }
-        finally
-        {
-            session?.Dispose();
-            connection.DecRef();
-        }
-    }
-
-    /// <summary>The identity response: who the client proved itself to be.</summary>
-    private static byte[] Identity(TlsSession session)
-    {
-        byte[] who = System.Text.Encoding.ASCII.GetBytes(session.PeerSubject ?? "anonymous");
-        byte[] head = System.Text.Encoding.ASCII.GetBytes(
-            $"HTTP/1.1 200 OK\r\ncontent-length: {who.Length}\r\n\r\n");
-
-        return [.. head, .. who];
-    }
 
     /// <summary>
     /// Whether the server DECLINED this client for a name. A drop counts, a timeout does not: a
@@ -648,50 +486,6 @@ internal static class RotationTests
         return outcome is Client.TlsOutcome.Refused or Client.TlsOutcome.Dropped;
     }
 
-    /// <summary>Completes the handshake and answers, so a certificate can be read AND used.</summary>
-    private static async Task OkHandler(Reactor reactor, TcpConnection connection)
-    {
-        TlsSession? session = null;
-
-        try
-        {
-            session = await reactor.GetService<TlsService>()!.AcceptAsync(connection);
-
-            // The first request routinely rides in with the handshake's final flight - a TLS 1.3
-            // client sends it immediately after Finished, and those bytes are already decrypted and
-            // gone from the socket by the time AcceptAsync returns. Answering it BEFORE parking on
-            // a read is what the send-first loop means; waiting first hangs, because the bytes
-            // being waited for already arrived.
-            if (!session.DrainPlaintext().IsEmpty)
-            {
-                session.Write(connection, Response);
-                await connection.FlushAsync();
-            }
-
-            // No ResetRead above: that belongs to a read that was actually issued, and calling it
-            // for one that never happened leaves the connection's read state describing a read
-            // nobody made.
-            while (true)
-            {
-                RecvSnapshot snapshot = await connection.ReadAsync();
-
-                if (snapshot.IsClosed)
-                {
-                    return;
-                }
-
-                session.Write(connection, "HTTP/1.1 200 OK\r\ncontent-length: 2\r\n\r\nok"u8.ToArray());
-
-                await connection.FlushAsync();
-                connection.ResetRead();
-            }
-        }
-        finally
-        {
-            session?.Dispose();
-            connection.DecRef();
-        }
-    }
 
     private static IEnumerable<(string Label, bool KernelTx)> Paths(bool ktls)
     {
